쪼금 수정

+++ Kirbies/01_WinMain/Camera.h
@@ -45,6 +45,6 @@ public:
 	inline float GetMoveSpeed()const { return mMoveSpeed; }
 	inline void SetMoveSpeed(float speed) { mMoveSpeed = speed; }
 	
-	POINT GetPoint(float x, float y)const { return { x - mRect.left, y - mRect.top }; }
+	POINT GetPoint(long x, long y)const { return { x - mRect.left, y - mRect.top }; }
 };
